# api/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workout Tracker Api",
    "version": "1.0.O",
    "description": "Api for Tracking workouts and progress"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "localserver"
    }
  ],
  "paths": {
...
    "/schedule/schedule": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scheduleDate": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/schedule": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scheduleDate": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}